feat(gis): move msg processing to gis controller
